% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign.R
\name{expDesign}
\alias{expDesign}
\title{Allocation of varieties to farmers}
\usage{
expDesign(nVarieties, nFarmers, method = c("alpha", "random"))
}
\arguments{
\item{nFarmers}{the number of farmers (observers), must be greater than or equal to nVarieties / 3}

\item{method}{must be one of "alpha" or "random"}

\item{nVarietiesieties}{the number of varieties}
}
\value{
an nFarmers * 3 matrix, with each row being the label of varieties assigned to that farmer
}
\description{
Generate experimental designs to allocate vartieties to farmers.
2 experimental designs are included, "alpha" and "random"
}
\examples{
#use alpha design
expDesign(120, 245, method = 'alpha')

}
