% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientDescent.R
\name{gradientDescent}
\alias{gradientDescent}
\title{Gradient descent function generalizable across ranking models}
\usage{
gradientDescent(data, targetFN = targetPL, sigma = diag(ncol(data)),
  rate = 0.01, maxiter = 1000, tol = 1e-08, startVar = 1,
  startScores = scale(rnorm(ncol(data))), decay = 0.2)
}
\arguments{
\item{data}{a n * m matrix,
where n is the number of observers and m is the number of items to rank;
each row vector is a partial ranking (triple comparisons),
with i-th element being the rank assigned to item i;
the entry where that item is not ranked in the partial ranking is replaced by 0}

\item{targetFN}{calculates the likelihood of the current score and the gradient
to move to the next iteration}

\item{sigma}{the covariance matrix among varieties}

\item{rate}{the learning rate}

\item{maxiter}{the maximum number of iterations}

\item{tol}{the tolerance}

\item{startVar}{initial guess at variance of scores}

\item{startScores}{initial guesses at scores}

\item{decay}{how fast the learning rate decays when likelihood doesn't increase}
}
\description{
Gradient descent function generalizable across ranking models
}
