% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankLM.R
\name{rankLM}
\alias{rankLM}
\title{Rank aggregation using linear models}
\usage{
rankLM(data, K = NULL)
}
\arguments{
\item{data}{a n * m matrix,
where n is the number of observers and m is the number of items each observer ranked;
each row vector is a partial or complete ranking,
with i-th element being the rank assigned to item i;
the entry where that item is not ranked in the partial ranking is replaced by 0}

\item{K}{the additive relationship matrix; if provided,
use to specify the covariance among item scores}
}
\value{
Return a list with three components:
    \item{ranks}{a vector where the i-th element is the rank assigned to the i-th item.}
    \item{scores}{a vector of the estimated scores of the varieties.}
    \item{scoreVar}{a scalar of the estimated score variance. This variance is method specific.}
}
\description{
Aggregate the ranking from triple comparisons into a consensus ranking using linear models.
}
