% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resNAD.R
\docType{data}
\name{resNAD}
\alias{resNAD}
\title{Simulation results for models without adherence parameters}
\format{Include the heritability, the model used (models with squared brackets use genotypic information), the number of varieties, the number of observations, the Kendall metric and the top20 metric}
\usage{
data(resNAD)
}
\description{
We do a simulation to see the performance of models
when the adherence parameters are not included.
Models without adherence parameters are simply recovered by taking adherences for all farmers to be 1.
We include LM, TH, BT and PL with and without genetypic information in this simulation.
We run the procedure 100 times with h^2 = 0.2, 0.5, 0.8 with nvar = 240 and nobs = 180, 240, 300, 360.
We use the alpha design to allocate varieties to farmers.
}
\examples{
data(resNAD)
summary(resNAD)
}
\keyword{datasets}

