% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetBT.R
\name{targetBT}
\alias{targetBT}
\title{Likelihood and gradient for the Bradley-Terry model}
\usage{
targetBT(scores, data, inv_sigma)
}
\arguments{
\item{scores}{the scores of the varieties. score[1] is the variance among scores}

\item{data}{a n * m matrix,
where n is the number of observers and m is the number of items to rank;
each row vector is a partial ranking (triple comparisons),
with i-th element being the rank assigned to item i;
if an item is not ranked its partial ranking is 0}

\item{inv_sigma}{the inverse of the score covariance matrix among varieties}
}
\description{
Likelihood and gradient for the Bradley-Terry model
}
