% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankAg.R
\name{rankAg}
\alias{rankAg}
\title{Ranking aggregation of triple comparisons}
\usage{
rankAg(data, K = diag(ncol(data)), method = "TH", rate = 0.01,
  maxiter = 1000, tol = 1e-08, startVar = 1, startScores = NULL,
  decay = 0.2, nTechRep = 2)
}
\arguments{
\item{data}{a n * m matrix,
where n is the number of observers and m is the number of items to rank;
each row vector is a partial ranking (triple comparisons),
with i-th element being the rank assigned to item i;
the entry where that item is not ranked in the partial ranking is replaced by 0}

\item{K}{the additive relationship matrix;
all methods must specify K except LM}

\item{method}{one of "BT", "PL", "TH", "LM"}

\item{rate}{the learning rate}

\item{maxiter}{the maximum number of iterations}

\item{tol}{the tolerance}

\item{startVar}{initial guess at variance of scores}

\item{startScores}{initial guesses at scores}

\item{decay}{how fast the learning rate decays when log post doesn't improve}

\item{nTechRep}{for methods that may not converge perfectly, try more than once and pick the best}
}
\value{
Return a list with three components:
    \item{ranks}{a vector where the i-th element is the rank assigned to the i-th item.}
    \item{scores}{a vector of the estimated scores of the varieties.}
    \item{scoreVar}{a scalar of the estimated score variance. This variance is method specific.}
}
\description{
Ranking aggregation using Bradley-Terry model, Plackett-Luce model, Thurstone model,
and linear model. This is the wrapper function for rankLM, sgdBT, sgdPL, sgdTH
}
\examples{
#synthetic ranking data
data <- rbind(c(3,1,2), c(3,1,2), c(3,2,1))
#use identity relationship matrix
K <- diag(1, 3)

#rank aggregation
rankAg(data, K, method = 'TH')

}
