% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankAg.R
\name{rankAg}
\alias{rankAg}
\title{Ranking aggregation of triple comparisons}
\usage{
rankAg(data, K = NA, method)
}
\arguments{
\item{data}{a n * m matrix,
where n is the number of observers and m is the number of items to rank;
each row vector is a partial ranking (triple comparisons),
with i-th element being the rank assined to item i;
the entry where that item is not ranked in the partial ranking is replaced by 0}

\item{K}{the additive relationship matrix;
all methods must specify K except LM}

\item{method}{one of "BT", "PL", "TH", "MPM", "LM"}
}
\value{
Return a list with two components:
    \item{ranks}{a vector where the i-th element is the rank assigned to the i-th item.}
    \item{ranking}{a vector where the i-th element is the item ranked in the i-th place}
}
\description{
Ranking aggregation using Bradley-Terry model, Plackett-Luce model, Thurstone model, Multinomial Preference model
and linear model. This is the wrapper function for rankLM, sgdBT, sgdPL, sgdThurs and sgdMPM.
}
\examples{
#synthetic ranking data
data = rbind(c(3,1,2), c(3,1,2), c(3,2,1))
#use identity relationship matrix
K = diag(1, 3)

#rank aggregation
rankAg(data, K, method = 'TH')




}
